Native to u32 conversion doesn't require additional swapping on big endian machines
On open_rom, the ROM is byte-swapped such that we get a "native" ROM (eg Z64).
Some hashes are then run, then the ROM is again byteswapped to be in u32
format (eg word-aligned accesses can be made directly without swapping).
This second swapping is not needed on Big Endian machines.